docs: Use new URL for github documentation
Github moved all documentation into a single place so the URL you point to is showing the message
`The content on this site may be out of date. For the most accurate and up-to-date content, visit https://docs.github.com/webhooks/creating/.`
When clicking no the link it redirects to the URL I updated in this commit.